requirements.txt.txt
streamlit
numpy
matplotlib
opencv-python
pillow
scikit-image
pandas
scipy
